Peale pikka jamamist transactionitega ja lisa fieldide lisamise otsustamist.

+++ KooliProjekt/Models/TransactionsViewModel.cs
@@ -9,11 +9,11 @@
         public int? AssetId { get; set; }
         public string? AssetName { get; set; }
         public string? Action { get; set; } // buy or sell
-        public double? BalanceBefore {  get; set; } // how much was in the wallet before
+        public decimal? BalanceBefore {  get; set; } // how much was in the wallet before
         public int? TransactedUnitAmount { get; set; } // how many assets were trasferred
-        public int? TransactionUnitCost {  get; set; } //how much one piece cost
-        public double? TransactionSum {  get; set; }
-        public double? BalanceAfter { get; set; }
+        public decimal? TransactionUnitCost {  get; set; } //how much one piece cost
+        public decimal? TransactionSum {  get; set; }
+        public decimal? BalanceAfter { get; set; }
         public string? AssetType { get; set; } // realestate or crypto or whatever
